updateToy page design updated

+++ src/Pages/MyToys/MyToys.jsx
@@ -1,10 +1,12 @@
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../Provider/AuthProvider";
 import MyToysData from "./MyToysData";
+import Swal from "sweetalert2";
 
 const MyToys = () => {
     const { user } = useContext(AuthContext);
     const [myToys, setMyToys] = useState([]);
+
     const url = `https://toy-store-server-blond.vercel.app/my-toys?email=${user?.email}`;
 
     useEffect(() => {
@@ -13,26 +15,42 @@ const MyToys = () => {
             .then((data) => setMyToys(data));
     }, [url]);
 
-    const handleDelete = (_id) => {
-        fetch(`https://toy-store-server-blond.vercel.app/toys/${_id}`, {
-            method: 'DELETE'
-        })
-        .then((res) => res.json())
-        .then((data) => {
-            if (data.deletedCount > 0) {
-                alert('Deleted');
-                const remainingToys = myToys.filter((toys) => toys._id !== _id);
-                setMyToys(remainingToys);
+    const handleDelete = _id => {
+        console.log(_id)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`https://toy-store-server-blond.vercel.app/toys/${_id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        if(data.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your product has been deleted.',
+                                'success'
+                            )
+                            setMyToys(prevMyToys => prevMyToys.filter(toy => toy._id !== _id));
+                        }
+                    })
             }
-        });
-    };
+        })
+    }
 
     return (
         <div className="container mx-auto p-4">
             <div className="overflow-x-auto">
-                <table className="min-w-full">
+                <table className="min-w-full text-center">
                     <thead>
-                        <tr className="bg-gray-700 text-white">
+                        <tr className="bg-gray-400 text-white">
                             <th className="p-3">ID</th>
                             <th className="p-3">Toy Name</th>
                             <th className="p-3">Sub-Category</th>
